backend/requirements.txt
fastapi>=0.104.1
sqlmodel>=0.0.8
uvicorn[standard]>=0.24.0
asyncpg>=0.29.0
psycopg2-binary
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0
alembic>=1.13.1
python-multipart>=0.0.6
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-jose[cryptography]>=3.3.0
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2
openai>=1.0.0
openai-agents>=0.8.0
aiosqlite>=0.19.0